Stop using public_items itself for tests
That makes it harder to rename the lib to `public-api`. Also do a big
restructuring on how rustdoc JSON is built for testing.

Building rustdoc JSON on-demand has several advantages over versioning
rustdoc JSON in-repo:
* Smaller repo size
* We don't need a special nightly job; it is sufficient to run regular CI nightly
* It is easier to manage updates to the rustdoc JSON format

Note that we keep some rustdoc JSON in-repo though, mainly so that
examples/ have something to use.