Rewrite diff algo with `hashbag` to fix off-by-one skewing
And add regression test called `no_off_by_one_diff_skewing()` that the
old algo does not pass.

This commit relies on a pending contribution to `hashbag`:
https://github.com/jonhoo/hashbag/pull/6

To not be blocked on that contribution, I ported that code into our own
repo. We can remove that code when/if that contribution is made
available in a new `hashbag` release.

Closes #50