Put `#[attributes]` in front of items
It is much better to take relevant attributes into account in SOME way
than in now way, even if the current solution is not ideal.

I experimented with rendering attributes on the line above the item in
question, but:
* diffing looks weird
* we need to special-case the render code and other code
* it does not follow the current "one line per item" philosophy

I consider it part of #37/#38 to present attributes in a more natural
manner.

But I think this change is enough to
Close #41