Add --deny=[all] flag
This patch adds an option `--deny=all` flag for checking whether the
diff is empty.

Right now only the "all" case is implemented, but more options such as
`--deny=deleted` or `--deny=added` might be implemented based on the
approach from this patch.

This patch was implemented with a lot of help and suggestions from
Martin.

Co-authored-by: Martin Nordholts <enselic@gmail.com>
Signed-off-by: Matthias Beyer <mail@beyermatthias.de>