IntermediatePublicItem: Remove `pub root: &'a Crate` field
It was kept around because in the future we might need it. But it makes
deriving `Debug` on `IntermediatePublicItem` awkward, because printing
one `IntermediatePublicItem` will list *all* items.

At this point it seems to just cause troubles, so let's remove it. We
can always add it back if needed. But it was not needed to inline
imports, so I think we will never need it.

And also finally derive `Debug`.