render.rs: Expand `parameterized_test!`; use regular code
I had to update tests for the new `rustdoc_types::Path` and was quite
confused with how the `parameterized_test!` macro functions, and
rust-analyzer was also confused, so it was hard to update the tests.

So stop using `parameterized_test!` and instead use regular code, which
IMHO makes the code clearer and easier to work with and modify.