Add `RenderingContext` to `render.rs`
This used to be just a regular arg that was passed around (see #78), but
was later removed. It turns out we need something like this after all.
Add it back, but instead of explicitly passing it around, pass it round
implicitly via `self` in a new `RenderingContext`. This makes it a bit
easier to work with.