Include Auto Trait Implementations in API listings (and diffs)
This is important, because it is easy to accidentally break semver
compatibility with subtle auto trait impl changes. Read
https://predr.ag/blog/toward-fearless-cargo-update/ for example.

I considered not showing these by default, and instead adding a
`--with-auto-trait-implementations` as an opt-in, but I think that
becomes too complicated. Especially for CI uses cases where you would
have to pass to many and too long flags to fully list and diff an API.

So instead, this PR always shows them. If it end up becoming too noisy,
my plan is to get rid of `--with-blanket-implementations` and instead
making the default output complete, and then add an opt-in to not
include blanket and auto trait impls, called something like `--terse` or
`--simplified`.

Also adjust how constant are rendered to make the auto trait impls
become rendered like in rustdoc HTML.