Add `PublicApi::from_rustdoc_json(...)` that takes rustdoc JSON file `Path`
Because in 95% of the use cases of this library, the user has a rustdoc
JSON file to work with rather than a `&str`.

I have refrained from adding this helper up until now, because I wanted
the API surface to be minimal, but I think this is a safe and
future-proof API addition by now.

Also make `from_rustdoc_json_str()` more generic while we're at it.