Move some tests from `test-invocation-variants.sh` to `cargo test`
Remove the `cargo run` test, because it is equivalent to running
`cargo-public-api`, which we already have plenty of tests for.

The other ones we port to `cargo test` after adding `./target/debug` to
`PATH` in the tests.

For me this makes `./scripts/run-ci-locally.sh` finish in ~12s instead
of ~16s.