Remove unnecessary ensure_impl_debug() test
I get a bit embarrassed every time I see this test. I wrote it very
early on in the project when I cared about code coverage a lot. This was
before we had e.g. `comprehensive_api` or the `public-api.txt` mechanism
via the `expect-test` crate.

So let's just remove it now.