Make `Eq` and `Hash` for `PublicItem` only take tokens into account
Otherwise there can be to users seemingly mysterious diffs, since the
`sortable_path` is not visible in the output. So it does not make sense
to take it into account when checking if two items are equal.

This means we also need to stop using `Ord` for sorting, so introduce
special-purposes `grouping_cmp` methods.