Add `--omit blanket-impls|auto-trait-impls|auto-derived-impls`
For day-to-day usage, the `--simplified` (`-s`) flag is still preferred.
However, when you paste a `cargo public-api` invocation together with
its output in e.g. a GitHub comment, the recipient of the comment is not
necessarily familiar with `cargo public-api` and its day-to-day flags.

By adding these explicit flags, we remove the need to explain what
`--simplified` means, because it becomes more or less obvious when these
more explicit forms are used.