public-api bin: Remove --print-minimum-nightly-rust-version
It was originally added for `scripts/test-invocation-variants.sh`, but
we have since then converted that script to regular `cargo test` tests,
so this flag is not needed any longer.

My intention is to eventually get rid of the public-api bin completely,
because I never use it, and I don't think anyone else does either.
`cargo public-api` is simply much better.

This commit is the first step on that journey.