Remove the `public-api` bin (NOT the `cargo-public-api` bin)
The `public-api` bin is a simple CLI wrapper around the `public_api`
library. It comes nowhere near the capabilities of `cargo public-api`.
So no one uses it.

Remove it.

It had a role to play in the early development of this project, but it's
not needed anymore.

The `cargo test`s that are removed in this commit are already covered by
other tests elsewhere in the project, or specific to the `public-api`
bin.