Make `diff-latest` feature mandatory
I think by now this feature is a corner-stone of cargo-public-api, and I
do not think it is worth the overhead to support cargo-public-api with
such a useful feature disabled.

Now is a good time to do this change since we need to bump 0.x.0 version
for other reasons anyway.