Fixup 'Avoid textual API diff when changing a trait impl to an auto-derived impl.'
An inherent impl can't be changed to an auto-derived impl. It is _trait_
impls that can be changed to auto-derived impls.

The regression test I added in
https://github.com/Enselic/cargo-public-api/pull/516 failed without the
(wrong) fix in https://github.com/Enselic/cargo-public-api/pull/516 and
passed with the (wrong) fix, but that was just an accident.

This is a fix of the fix. Honestly I'm too lazy to come up with a
regression test for the regression test. Clearly the fix was logically
wrong in the first place, so I will re-use the regression test I already
added also for this fix. I have confirmed that the regression test fails
without the (correct) fix.