Allow `--package`, `--color` etc also after subcommands such as `diff`
Before, the following would not work, because the `--package` and
`--color` options were not allowed after subcommands such as `diff`

    cargo public-api diff --package foo --color always

Now that command works. The help texts becomes more noisy, but I think
that is a fair price.

I added a couple of regression tests that fails without the fix, but I
did not bother to be exhaustive.