Include pub const types in more cases such as for arrays and tuples
In general, Rust semver is about backwards compatibility on the source
code level. And changing a value of a constant does not break source
code compatibility. Dependent crates will still build. So we should not
include values as part of the API surface.

We did have the wrong logic with regards to rendering types though. Fix
that.