README.md: Don't reference CHANGELOG.md
I like READMEs to be as simple as possible, and see references to
CHANGELOG.md as noise. If someone is looking for a changelog it is very
easy to find the right one.

And let's be honest, very few people read the changelog. So it does not
make sense to take up valuable space in the README for it.