Auto-merge-dependabot-PRs.yml: Use --repo instead of checkout
`pull_request_target` is the most sensitive kind
of trigger because it is the most prevalent target
of own requests (see e.g.
https://blog.yossarian.net/2024/12/06/zizmor-ultralytics-injection)

So harden our only `pull_request_target` job a bit
by removing `actions/checkout`. That means we
don't have to worry about the
`persist-credentials` option (which is `true` by
default).