Auto-merge-dependabot-PRs.yml: Do not evaulate context expressions in shell context
Auto-merge-dependabot-PRs.yml is our only
`pull_request_target` triggered workflow and that
type is extra risky.

From
https://securitylab.github.com/resources/github-actions-untrusted-input/
we can read:

> The best practice to avoid code and command
> injection vulnerabilities in GitHub workflows is
> to set the untrusted input value of the
> expression to an intermediate environment
> variable

And even though `${{ github.event.pull_request.number }}`
is not free-text user controlled, there might be a
bug in GitHub itself that allows free text in that
field today or in the future. So protect us from
pwn requests even in that unlikely scenario by
putting the PR number in an env var.