CI: Stop testing with CARGO_REGISTRIES_CRATES_IO_PROTOCOL=git
Ubuntu with

    CARGO_REGISTRIES_CRATES_IO_PROTOCOL=git

takes 6 minutes and slows down CI. With

    CARGO_REGISTRIES_CRATES_IO_PROTOCOL=sparse

it just takes 2 minutes.

sparse became the default more than 1,5 years ago
and I doubt anyone is still using it that is also
using cargo-public-api:
https://blog.rust-lang.org/2023/06/01/Rust-1.70.0.html#sparse-by-default-for-cratesio

But keep the code that makes it work around still,
since that doesn't really hurt us. But if we start
to get problems with it, like regressions, we
remove that code instead of fix it..