Remove Auto-merge-dependabot-PRs.yml, we will use a GitHub App instead
We will use my app https://github.com/apps/auto-merge-dependabot-prs
which I still need to apply finishing touches to but I'm confident
enough in it that I want to start using it. It is already activated in
this repo.

I'll probably manually have to auto-merge some dependabot PRs during a
transition period.